requirements.txt
dnslib==0.9.23
dnspython==2.4.2
scapy==2.5.0
